+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./gulp-utils";
 export * from "./lint";
 export * from "./publish";
 export * from "./transpile";
+export * from "./prepublish";