+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.7.2](https://github.com/sketch7/ssv-tools/compare/0.7.1...0.7.2) (2018-10-25)
+
+### Features
+
+- **prepublish:** implement `prepublish` which invokes `writePackageTransform` + copy contents such as `CHANGELOG.md`, `LICENSE` etc...
+- **prepublish:** implement `writePackageTransform`
+
+
 ## [0.7.1](https://github.com/sketch7/ssv-tools/compare/0.7.0...0.7.1) (2018-10-25)
 
 ### Bug Fixes